pyproject.toml
[tool.poetry]
name = "gcaltools"
version = "1.0.0"
description = ""
authors = ["Steve De Jongh <dejongh.st@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
pyyaml = "^6.0"
rich = "^13.4.2"
xlsxwriter = "^3.1.2"
pytz = "^2023.3"
oauth2client = "^4.1.3"
google-api-python-client = "^2.92.0"
prompt-toolkit = "^3.0.39"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
gcaltools = "gcaltools.gcaltools:main"
